transition to dashboard; set all pairs and current pair
